// src/shared/types/user.ts
//интерфейс пользователя

export enum UserType {
  Normal = 'normal',
  Pro = 'pro',
}

export interface IUser {
  name: string; // Имя пользователя, обязательное поле, длина 1-15 символов
  email: string; // Электронная почта, обязательное поле, должен быть валидным email
  avatar?: string; // Аватар пользователя, необязательное поле, ссылка на изображение .jpg или .png
  userType: UserType; // Тип пользователя, обязательное поле, возможные значения: 'normal', 'pro', может потом изменю на булевое значение где 1 - pro
}